Clean up and refactor Streak functions
