+++ src/components/DailyHabitCard/DailyHabitCard.jsx
@@ -3,13 +3,8 @@ import useStyles from "./useStyles";
 import { Grid, Container, Paper, Typography } from "@material-ui/core";
 import { Context } from "../../Store/habitStore";
 import { makeStyles } from "@material-ui/core/styles";
-import {
-  getDay,
-  parse,
-  differenceInDays,
-  isYesterday,
-  isToday,
-} from "date-fns";
+import { getDay } from "date-fns";
+import { calculateCurrentStreak } from "../../utilities/calculateStreak";
 
 function DailyHabitCard({ habit }) {
   const classes = useStyles();
@@ -34,16 +29,6 @@ function DailyHabitCard({ habit }) {
   const dateColor = {
     color: colors[habit.color],
   };
-
-  //  const demoData = [
-  //    { date: "11112021", isComplete: true },
-  //    { date: "10112021", isComplete: true },
-  //    { date: "08112021", isComplete: true },
-  //    { date: "12102021", isComplete: true },
-  //    { date: "11102021", isComplete: true },
-  //    { date: "09102021", isComplete: true },
-  //    { date: "06102021", isComplete: true },
-  //  ];
 
   useEffect(() => {
     if (state.habits) {
@@ -191,145 +176,6 @@ function DailyHabitCard({ habit }) {
   return state.habitStatus[habit.habitName][todayDayIndex]
     ? SelectedDate(todayDayIndex)
     : UnSelectedDate(todayDayIndex);
-}
-
-function calculateLongestStreak(habit) {
-  const sortedDates = sortByDates(habit, "asc");
-  let result = sortedDates.reduce(
-    (streak, habit) => {
-      if (streak.previousDate === null) {
-        streak.previousDate = habit.date;
-      } else {
-        if (habit.isComplete) {
-          let preDate = parse(streak.previousDate, "ddMMyyyy", new Date());
-          let currDate = parse(habit.date, "ddMMyyyy", new Date());
-          let diff = differenceInDays(currDate, preDate);
-          if (diff === 1) {
-            if (streak.tempStartDate === null) {
-              streak.tempStartDate = preDate;
-            }
-
-            streak.count = streak.count + 1;
-            streak.longestStreak = Math.max(streak.count, streak.longestStreak);
-            if (streak.count >= streak.longestStreak) {
-              streak.startDate = streak.tempStartDate;
-              streak.endDate = currDate;
-            }
-          } else {
-            if (streak.count >= streak.longestStreak) {
-              streak.endDate = preDate;
-            }
-            streak.tempStartDate = null;
-
-            streak.count = 1;
-          }
-
-          streak.previousDate = habit.date;
-        }
-      }
-      return streak;
-    },
-    {
-      count: 1,
-      longestStreak: 0,
-      previousDate: null,
-      tempStartDate: null,
-      tempEndDate: null,
-      startDate: null,
-      endDate: null,
-    }
-  );
-  return result;
-}
-
-function calculateCurrentStreak(habit) {
-  // sorted the habit tracker date in descending order
-  const sortedDates = sortByDates(habit, "desc");
-
-  // check if the latest date of habit tracked data is yesterday
-  const isLatestDateIsYesterday = isYesterday(
-    parse(sortedDates[0].date, "ddMMyyyy", new Date())
-  );
-
-  // check if the latest date of habit tracked data is today
-  const isLatestDateIsToday = isToday(
-    parse(sortedDates[0].date, "ddMMyyyy", new Date())
-  );
-
-  if (isLatestDateIsYesterday || isLatestDateIsToday) {
-    // run reducer to calculate the count of the continues streak
-    return calculateRecentContinuousStreak(sortedDates);
-  }
-
-  if (!isLatestDateIsToday && !isLatestDateIsYesterday) {
-    return {
-      currentStreak: 0,
-      startDate: null,
-      endDate: null,
-    };
-  }
-
-  // if habitTrack date array does not have yesterday date then current Streak is  0
-
-  // else  if it has yesterday date date then caculate the longest stresk form yesterday.
-}
-
-function calculateRecentContinuousStreak(trackData) {
-  const result = trackData.reduce(
-    (streak, trackData) => {
-      if (streak.previousDate === null) {
-        streak.previousDate = trackData.date;
-      } else {
-        if (trackData.isComplete) {
-          let preDate = parse(streak.previousDate, "ddMMyyyy", new Date());
-          let currDate = parse(trackData.date, "ddMMyyyy", new Date());
-          let diff = differenceInDays(preDate, currDate);
-          if (diff === 1) {
-            if (streak.tempStartDate === null) {
-              streak.tempStartDate = preDate;
-            }
-            streak.count += 1;
-            streak.currentStreak = Math.max(streak.count, streak.currentStreak);
-            streak.startDate = currDate;
-            streak.endDate = streak.tempStartDate;
-          } else {
-            return streak;
-          }
-          streak.previousDate = trackData.date;
-        }
-      }
-      return streak;
-    },
-    {
-      count: 1,
-      tempStartDate: null,
-      previousDate: null,
-      currentStreak: 1,
-      startDate: null,
-      endDate: null,
-    }
-  );
-
-  return result;
-}
-
-function sortByDates(habit, option) {
-  if (option === "asc") {
-    const res = habit.sort(
-      (a, b) =>
-        parse(a.date, "ddMMyyyy", new Date()) -
-        parse(b.date, "ddMMyyyy", new Date())
-    );
-    return res;
-  }
-  if (option === "desc") {
-    const res = habit.sort(
-      (a, b) =>
-        parse(b.date, "ddMMyyyy", new Date()) -
-        parse(a.date, "ddMMyyyy", new Date())
-    );
-    return res;
-  }
 }
 
 export default DailyHabitCard;
